Fix deleting selected chat

+++ domain/src/main/java/ru/kima/intelligentchat/domain/chat/useCase/DeleteChatUseCase.kt
@@ -1,10 +1,12 @@
 package ru.kima.intelligentchat.domain.chat.useCase
 
 import kotlinx.coroutines.flow.first
+import ru.kima.intelligentchat.domain.card.repository.CharacterCardRepository
 import ru.kima.intelligentchat.domain.chat.repository.ChatRepository
 
 class DeleteChatUseCase(
     private val chatRepository: ChatRepository,
+    private val characterCardRepository: CharacterCardRepository,
     private val subscribeToFullChatUseCase: SubscribeToFullChatUseCase,
 ) {
     suspend operator fun invoke(chatId: Long) {
@@ -14,6 +16,9 @@ class DeleteChatUseCase(
             else -> return
         }
 
-        chatRepository.deleteChat(fullChat)
+        val card = characterCardRepository.getCharacterCard(fullChat.cardId).first()
+        if (card.selectedChat != chatId) {
+            chatRepository.deleteChat(fullChat)
+        }
     }
 }
